Move pbench libraries to its own namespace
This commit does several things at once:

- Create the pbench for both sever and agent so that
  in the future both agent and server can share code.
- Move the python pbench-server python libraries to
  the new pbench.sever nameapce.
- Move the s3backup class to the new pbench.server
  namespace.
- Move the configtools class to pbench.lib since
  both agent and server both use this class.
- Adjust module imports to take account the new
  pbench namespace.
- Adjust the unittest shell scripts to take
  in account that the pbench/lib/configtool
  directory no longer exists.
