+++ .pre-commit-config.yaml
@@ -3,7 +3,6 @@ repos:
     rev: 3.7.9
     hooks:
      - id: flake8
-       additional_dependencies: [flake8-typing-imports==1.6.0]
   - repo: https://github.com/python/black.git
     rev: 19.10b0
     hooks:
