+++ pbench/server/__init__.py
@@ -5,7 +5,6 @@ Simple module level convenience functions.
 import sys
 import os
 import logging
-import configtools
 import shutil
 import hashlib
 
@@ -14,6 +13,8 @@ from time import time as _time
 from datetime import datetime, tzinfo, timedelta
 from functools import partial
 from configparser import ConfigParser, NoSectionError, NoOptionError
+
+from pbench.lib import configtools
 
 
 class simple_utc(tzinfo):
