+++ server/bin/pbench-index.py
@@ -14,8 +14,8 @@ import tempfile
 from argparse import ArgumentParser
 from configparser import Error as ConfigParserError
 
-from pbench import tstos, BadConfig, quarantine, rename_tb_link
-from pbench.indexer import (
+from pbench.server import tstos, BadConfig, quarantine, rename_tb_link
+from pbench.server.indexer import (
     IdxContext,
     ConfigFileError,
     BadDate,
@@ -30,7 +30,7 @@ from pbench.indexer import (
     TemplateError,
     VERSION,
 )
-from pbench.report import Report
+from pbench.server.report import Report
 
 
 # Internal debugging flag.
