+++ server/bin/unittests
@@ -240,8 +240,6 @@ function _setup_state {
     let res=res+$?
     mkdir -p $_testopt_sat/bin
     let res=res+$?
-    mkdir -p $_testopt/common/lib
-    let res=res+$?
     cp -a $_tdir/{unittests,getconf.py,pbench*} $_testopt/bin
     let res=res+$?
     mkdir -p $_testopt/html/static/{js,css}/v0.{2,3}
@@ -255,8 +253,6 @@ function _setup_state {
     cp -a $_tdir/../../web-server/v0.3/css/ $_testopt/html/static/css/v0.3
     let res=res+$?
     cp -a $_tdir/{unittests,getconf.py,pbench*} $_testopt_sat/bin
-    let res=res+$?
-    cp -a $_tdir/../../lib/* $_testopt/common/lib
     let res=res+$?
     cp -a $_tdir/../lib $_testopt
     let res=res+$?
